+++ src/enum.ts
@@ -1,5 +1,7 @@
-export const generateMockID = "mockServer.generateMockFromHAR";
-export const startServerID = "mockServer.startServer";
-export const stopServerID = "mockServer.stopServer";
-export const resetServerId = "mockServer.resetServer";
-export const switchEnvironmentID = "mockServer.switchEnvironment";
+export const GENERATE_ROUTES = "mockServer.generateRoutes";
+export const START_SERVER = "mockServer.startServer";
+export const STOP_SERVER = "mockServer.stopServer";
+export const RESET_SERVER = "mockServer.resetServer";
+export const SWITCH_ENVIRONMENT = "mockServer.switchEnvironment";
+export const GET_DB_SNAPSHOT = "mockServer.getDbSnapShot";
+export const HOMEPAGE = "mockServer.homePage";